# public/openapi.yaml
openapi: 3.0.3
info:
  title: 同频 B2 API
  version: 1.3.0
  description: 同频 B2 情绪急救与数字陪伴前后端接口。支持 Swagger UI Try it out；可在 Servers 中切换本地或远程环境。注意：本 API 不替代专业医疗或心理服务，AI 输出必须人工确认。
servers:
- url: http://localhost:8080/api/v1
  description: 本地开发环境
- url: https://{remoteHost}/api/v1
  description: 远程环境（修改 remoteHost 为实际域名）
  variables:
    remoteHost:
      default: api.tongpin.example.com
      description: 远程 API 域名，不带 https:// 和路径
...
paths:
...
  /manager/ai/config:
    get:
      tags:
      - AI 配置
      summary: 获取当前 AI 配置
      operationId: getAIConfig
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIConfig'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - AI 配置
      summary: 创建 AI 配置草稿
      operationId: createAIConfig
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              provider: deepseek
              base_url: https://api.deepseek.com
              model: deepseek-v4-flash
              api_key_secret_ref: deepseek_prod_key
              enabled: true
              default_parameters:
                thinking:
                  type: disabled
                reasoning_effort: low
                max_tokens: 800
                temperature: 0.2
                top_p: 0.9
                response_format:
                  type: json_object
                stream: false
              timeout_ms: 2000
              max_retries: 1
              daily_token_budget: 200000
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIConfig'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/ai/config/{id}/publish:
    post:
      tags:
      - AI 配置
      summary: 发布 AI 配置
      operationId: publishAIConfig
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: id 标识
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              confirm_no_sensitive_data_in_test: true
              safety_review_passed: true
              change_note: 发布 v4 配置。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIConfig'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/ai/config/{id}/test:
    post:
      tags:
      - AI 配置
      summary: 使用合成输入测试 AI 配置
      operationId: testAIConfig
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: id 标识
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              task_type: chat_suggestion
              synthetic_input:
                age_band: 14-17
                main_request: 希望有人听我说说
                current_safety: safe
              execute_tools: false
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/ai/prompts:
    get:
      tags:
      - AI 配置
      summary: 获取提示词库
      operationId: listAIPrompts
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - AI 配置
      summary: 新增提示词版本
      operationId: createAIPrompt
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              task_type: chat_suggestion
              name: 倾听建议提示词
              system_prompt: 你是志愿者倾听辅助助手。不能诊断，不能建议用药。
              user_template: 当前诉求：{{main_request}}；当前安全：{{current_safety}}。请输出 JSON。
              variables:
              - main_request
              - current_safety
              - age_band
              safety_constraints:
              - no_diagnosis
              - no_medication_advice
              - no_external_action
              - requires_human_confirmation
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/ai/prompts/{id}:
    patch:
      tags:
      - AI 配置
      summary: 启用、停用或修改提示词草稿
      operationId: updateAIPrompt
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: id 标识
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              status: draft
              system_prompt: 更新后的安全提示词。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /manager/ai/skills:
    get:
      tags:
      - AI 配置
      summary: 获取技能库
      operationId: listAISkills
      description: ''
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - AI 配置
      summary: 新增技能版本
      operationId: createAISkill
      description: ''
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
            example:
              code: safety_gate
              name: 安全闸门
              description: 识别需要立即转人工或现实求助的情境。
              instructions: 只要出现当前不安全、身体急症或迫近危险，输出 professional_handoff。
              allowed_task_types:
              - chat_suggestion
              forbidden_actions:
              - diagnosis
              - external_call
              - notify_contact
              enabled: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 输入登录后获取的 access token。匿名用户先调用 /auth/user-session。
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: 写接口幂等键。重复提交同一请求时应返回同一结果。
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
...
    AIConfig:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
        status:
          type: string
        provider:
          type: string
        base_url:
          type: string
        model:
          type: string
        api_key_configured:
          type: boolean
        enabled:
          type: boolean
        default_parameters:
          type: object
          additionalProperties: true
        task_overrides:
          type: object
          additionalProperties: true
        prompt_ids:
          type: object
          additionalProperties: true
        skill_ids:
          type: array
          items:
            type: string
        tool_ids:
          type: array
          items:
            type: string
        timeout_ms:
          type: integer
        max_retries:
          type: integer
        daily_token_budget:
          type: integer
        published_at:
          type:
          - string
          - 'null'
          format: date-time
...
    GenericObject:
      type: object
      additionalProperties: true